tests/myproject/requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
numpy==2.2.0
pandas==2.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
slack_sdk==3.33.5
sqlparse==0.5.2
tzdata==2024.2
